🧬 Extract the common object for capturing output value as a common module.
